requirements.txt
requests
pandas
scikit-learn
Snakelike
os
matplotlib.pyplot
seaborn
hashlib
numpy